+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/LookAndFeelSelectionDialogBuilder.java
@@ -18,6 +18,8 @@
  */
 package is.codion.swing.common.ui.dialog;
 
+import is.codion.common.Configuration;
+import is.codion.common.property.PropertyValue;
 import is.codion.swing.common.ui.control.Control;
 import is.codion.swing.common.ui.laf.LookAndFeelEnabler;
 
@@ -30,6 +32,18 @@ import java.util.function.Consumer;
 public interface LookAndFeelSelectionDialogBuilder {
 
 	/**
+	 * <p>Specifies whether to include the platform look and feels in the selection combo box by default, if auxiliary ones are provided.
+	 * <p>Note that this has no effect if only the platform look and feels are provided.
+	 * <ul>
+	 * <li>Value type: Boolean
+	 * <li>Default value: false
+	 * </ul>
+	 * @see is.codion.swing.common.ui.laf.LookAndFeelProvider
+	 */
+	PropertyValue<Boolean> INCLUDE_PLATFORM_LOOK_AND_FEELS =
+					Configuration.booleanValue(LookAndFeelSelectionDialogBuilder.class.getName() + ".includePlatformLookAndFeels", false);
+
+	/**
 	 * @param owner the dialog owner
 	 * @return this builder
 	 */
@@ -40,6 +54,13 @@ public interface LookAndFeelSelectionDialogBuilder {
 	 * @return this builder
 	 */
 	LookAndFeelSelectionDialogBuilder enableOnSelection(boolean enableOnSelection);
+
+	/**
+	 * @param includePlatformLookAndFeels true if the platform look and feels should be included by default in the selection combo box
+	 * @return this builder
+	 * @see #INCLUDE_PLATFORM_LOOK_AND_FEELS
+	 */
+	LookAndFeelSelectionDialogBuilder includePlatformLookAndFeels(boolean includePlatformLookAndFeels);
 
 	/**
 	 * Displays a dialog allowing the user the select between all available Look and Feels.