+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/laf/PlatformLookAndFeelProvider.java
@@ -1,0 +1,57 @@
+/*
+ * This file is part of Codion.
+ *
+ * Codion is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Codion is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Codion.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * Copyright (c) 2025, Björn Darri Sigurðsson.
+ */
+package is.codion.swing.common.ui.laf;
+
+import javax.swing.UIManager;
+import javax.swing.UIManager.LookAndFeelInfo;
+import java.util.Collection;
+import java.util.stream.Stream;
+
+import static java.util.stream.Collectors.toList;
+
+/**
+ * <p>Provides the built-in platform look and feels.
+ * <p>Note that Nimbus and Motif are excluded.
+ */
+public final class PlatformLookAndFeelProvider implements LookAndFeelProvider {
+
+	@Override
+	public Collection<LookAndFeelEnabler> get() {
+		return Stream.of(UIManager.getInstalledLookAndFeels())
+						.filter(PlatformLookAndFeelProvider::included)
+						.map(PlatformLookAndFeelEnabler::new)
+						.collect(toList());
+	}
+
+	private static boolean included(LookAndFeelInfo lookAndFeelInfo) {
+		return !lookAndFeelInfo.getClassName().contains("Motif") && !lookAndFeelInfo.getName().contains("Nimbus");
+	}
+
+	private static final class PlatformLookAndFeelEnabler extends DefaultLookAndFeelEnabler {
+
+		private PlatformLookAndFeelEnabler(LookAndFeelInfo lookAndFeelInfo) {
+			super(lookAndFeelInfo);
+		}
+
+		@Override
+		public boolean platform() {
+			return true;
+		}
+	}
+}